factored out loading animation
